+++ src/test/kotlin/moe/gogo/lang/ast/DeclareStmntTest.kt
@@ -5,6 +5,7 @@ import io.kotlintest.specs.StringSpec
 import moe.gogo.lang.env
 import moe.gogo.lang.id
 import moe.gogo.lang.num
+import moe.gogo.lang.type.int
 
 class DeclareStmntTest : StringSpec() {
 
@@ -16,7 +17,7 @@ class DeclareStmntTest : StringSpec() {
             DeclareStmnt(listOf(id("var"), id("b"), id("="), num(10))).eval(e2)
 
             e1.contains("a") shouldBe true
-            e2["b"] shouldBe 10
+            e2["b"].int() shouldBe 10
         }
     }
