+++ src/test/kotlin/moe/gogo/lang/ast/StatementTest.kt
@@ -5,12 +5,13 @@ import io.kotlintest.specs.StringSpec
 import moe.gogo.lang.env
 import moe.gogo.lang.eval
 import moe.gogo.lang.num
+import moe.gogo.lang.type.int
 
 class StatementTest : StringSpec() {
 
     init {
         "eval"{
-            Statement(listOf(num(78))).eval() shouldBe 78
+            Statement(listOf(num(78))).eval().int() shouldBe 78
         }
     }
