+++ src/test/kotlin/moe/gogo/lang/ast/StatementsTest.kt
@@ -6,12 +6,13 @@ import moe.gogo.lang.ast.Statements.Companion.statements
 import moe.gogo.lang.env
 import moe.gogo.lang.eval
 import moe.gogo.lang.num
+import moe.gogo.lang.type.int
 
 class StatementsTest : StringSpec() {
 
     init {
         "single statement"{
-            statements(listOf(num(3))).eval() shouldBe 3
+            statements(listOf(num(3))).eval().int() shouldBe 3
         }
         "multi statements"{
             val s1 = statements(listOf(num(6)))
@@ -19,7 +20,7 @@ class StatementsTest : StringSpec() {
             val s3 = statements(listOf(num(4),s2))
             s3 as Statements
             s3.statements.size shouldBe 3
-            s3.eval() shouldBe 6
+            s3.eval().int() shouldBe 6
         }
     }
